% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_Bay_CV_SE_jackknife_individual.R
\name{estimate_Bay_CV_SE_jackknife_individual}
\alias{estimate_Bay_CV_SE_jackknife_individual}
\title{Jackknife estimator (individuals deletion) of Standard Errors of  Between-Group Effect for 2-level latent variable model with control variables estimated with regularized_Bayesian_estimator}
\usage{
estimate_Bay_CV_SE_jackknife_individual(data)
}
\arguments{
\item{data}{A list containing the dataset information with elements:
\itemize{
  \item `n`: Number of individuals per group.
  \item `k`: Number of groups.
  \item `x`: Predictor variable.
  \item `y`: Response variable.
  \item `C`: Control variables matrix (if applicable).
  \item `kc`: Number of control variables (1 or more).
}}
}
\value{
A list containing:
  \itemize{
    \item `beta_b_Bay`: The original Bayesian estimate of the between-group effect.
    \item `beta_b_Bay_ML`: The original Maximum Likelihood estimate of the between-group effect.
    \item `SE_beta_Bay_jackknife_individual`: The jackknife standard error for the Bayesian estimator.
    \item `SE_beta_Bay_ML_jackknife_individual`: The jackknife standard error for the Maximum Likelihood estimator.
  }
}
\description{
This function performs a jackknife estimation of standard errors by individually deleting one observation
from each group in a two-level latent variable model with control variables. The model uses a regularized
Bayesian estimator to compute between-group effects. For each jackknife replication, one individual is 
randomly deleted from each group, and the model is re-estimated to provide an estimation of the standard errors 
through the jackknife technique.
}
\details{
Two types of standard errors are calculated:
  - `SE_beta_Bay_jackknife_individual`: The jackknife standard error for the Bayesian estimator.
  - `SE_beta_Bay_ML_jackknife_individual`: The jackknife standard error for the Maximum Likelihood estimator.
}
\examples{
# Assuming data is a list structured as per the requirements
result_Bay_CV_SE_jackknife_individual <- estimate_Bay_CV_SE_jackknife_individual(data)
print(result_Bay_CV_SE_jackknife_individual)
}
