% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlob.R
\name{mlob}
\alias{mlob}
\title{Multi-Level Optimal Bayes Function (MLOB)}
\usage{
mlob(
  formula,
  data,
  group,
  balancing.limit = 0.2,
  conf.level = 0.05,
  jackknife = FALSE,
  punish.coeff = 2,
  ...
)
}
\arguments{
\item{formula}{A formula specifying the model (e.g., \code{Y ~ X + C...}) where Y is the dependent variable, X is the context variable,which is the focus of most applications of the model  (always included), and C...are any additional covariates.}

\item{data}{A data frame containing the variables referenced in the formula.All variables used in the model,including the dependent variable,context variable and covariates,must be present in this data frame.}

\item{group}{A positive numeric value representing the number of groups (J) in the data. The number of rows in the data must be divisible by group.}

\item{balancing.limit}{A number that represents the threshold of the maximum relative  part of data set from when the balancing takes place.Default to \code{0.2}}

\item{conf.level}{A numeric value representing the confidence level used to calculate confidence intervals for the estimators. Defaults to \code{0.95},corresponding to a 95\% confidence level.}

\item{jackknife}{Logical. If \code{TRUE}, the jackknife re-sampling method will be applied to  calculate the standard error of between-group and its confidence interval coefficient. Defaults to \code{FALSE}.}

\item{punish.coeff}{A multiplier that increases the number of items deleted from groups smaller than the target size, adding a penalty for undersized groups. Higher values intensify the penalty.Defaults to \code{2}.}

\item{...}{Additional arguments passed to the function.}
}
\value{
A list containing the results of the regularized Bayesian estimation, which includes the model formula,dependent and context variables,and other relevant details from the analysis.
}
\description{
MultiLevelOptimalBayes (MLOB) is specifically designed for small-sample, two-level latent variable models, which are frequently used in fields such as psychology, education, and other disciplines that involve hierarchical data structures.
This function estimates the between-group coefficient beta_b in a  multilevel latent variable model with covariates,using regularized Bayesian estimator.
It is designed for scenarios  where data is organized into multiple groups and requires that the data are balanced across these groups.
}
\details{
This function also verifies  whether the data is balanced (i.e., whether each group contains the same number of individuals).If the data is unbalanced,the balancing.limit argument
comes into effect, and identifies the optimal number of items and groups to delete based on the punishment criteria.
}
\examples{
# Example usage with the iris dataset


# View summary statistics (similar to summary of a linear model)
summary(result)
mlob(Sepal.Length ~ Sepal.Width + Petal.Length, data = data, group = 3, conf.level = 0.01, jackknife = FALSE)

Summary of Coefficients:
Estimate Std. Error Lower CI (99\%) Upper CI (99\%)   Z value   Pr(>|z|) Significance
beta_b             0.4279681  0.7544766     -1.5154349       2.371371 0.5672384 0.57055223
gamma_Petal.Length 0.4679522  0.2582579     -0.1972762       1.133181 1.8119567 0.06999289            .
For comparison, summary of coefficients from unoptimized analysis (ML):
 Estimate   Std. Error Lower CI (99\%) Upper CI (99\%)      Z value   Pr(>|z|) Significance
beta_b             0.6027440 5.424780e+15  -1.397331e+16   1.397331e+16 1.111094e-16 1.00000000
gamma_Petal.Length 0.4679522 2.582579e-01  -1.972762e-01   1.133181e+00 1.811957e+00 0.06999289            .

Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1


}
